ptr to fun creating shape in Shape class
Shapes are still being created by ShapeFactory, but a pointer to appropriate
function is stored in Shape class. Is ensures that every invocation of
initClass outside ShapeFactory (which is possible) doesn't make ShapeFactory
create outdated shapes - it is theoretically possible that shape class is
using different methods for creating shapes based on initClass parameters.